[LAK-3357 | PM-JS] seperated kafka unit with and without ssl and moved common functionality to abstract class, added script to generate certificates, added more integration tests for kafka ssl
