+++ build.gradle
@@ -54,11 +54,11 @@ uploadArchives {
         mavenDeployer {
             beforeDeployment { MavenDeployment deployment -> signing.signPom(deployment) }
 
-            repository(url: "https://services.api.bskyb.com/nexus/content/repositories/sky-releases/") {
+            repository(url: skyReleasesRepo) {
                 authentication(userName: archivaUser, password: archivaPass)
             }
 
-            snapshotRepository(url: "https://services.api.bskyb.com/nexus/content/repositories/sky-snapshots/") {
+            snapshotRepository(url: skySnapshotsRepo) {
                 authentication(userName: archivaUser, password: archivaPass)
             }
         }
