--- create-certs.sh
@@ -1,13 +1,0 @@
-#!/bin/bash
-#Step 1
-keytool -keystore server.keystore.jks -alias localhost -validity 365 -keyalg RSA -genkey
-#Step 2
-openssl req -new -x509 -keyout ca-key -out ca-cert -days 365
-keytool -keystore server.truststore.jks -alias CARoot -import -file ca-cert
-keytool -keystore client.truststore.jks -alias CARoot -import -file ca-cert
-#Step 3 create server keystore
-keytool -keystore server.keystore.jks -alias localhost -certreq -file cert-file
-openssl x509 -req -CA ca-cert -CAkey ca-key -in cert-file -out cert-signed -days 365 -CAcreateserial -passin pass:test1234
-keytool -keystore server.keystore.jks -alias CARoot -import -file ca-cert
-keytool -keystore server.keystore.jks -alias localhost -import -file cert-signed
-
